Change numbering for color patterns
This renumbering assigns lower numbers to the patterns with 48 edges and
higher numbers to the patterns with 50 edges, because this seems to be
the original ordering of the bottom sides for the inner pieces of the
Eternity II puzzle when they are ordered by piece number. It also seems
to be the natural ordering for the clue puzzles 1 and 2, such that their
used colors have consecutive numbers.